Remove separate `closures` module
