--- crates/bevy_reflect/src/func/closures/mod.rs
@@ -1,9 +1,0 @@
-pub use dynamic_closure::*;
-pub use dynamic_closure_mut::*;
-pub use into_closure::*;
-pub use into_closure_mut::*;
-
-mod dynamic_closure;
-mod dynamic_closure_mut;
-mod into_closure;
-mod into_closure_mut;
